Add some test
